Update dependencies to latest versions
The dependencies and dev dependencies have been updated to the latest
versions available as of 2016-05-18. This includes replacing
grunt-jasmine-node (no longer maintained) with grunt-jasmine-nodejs.
Updates to the test specs and spec runners were necessary to ensure that
the specs run properly in the latest version of Jasmine and in the IGB.
The previous setup of bundling the test specs was removed to help with
spec troubleshooting in the IGB.